Consolidate imports and text types
