+++ tools/mc-dart/src/main/java/io/spine/tools/mc/dart/gradle/McDartPlugin.java
@@ -31,7 +31,7 @@ import io.spine.tools.gradle.ProtoDartTaskName;
 import io.spine.tools.gradle.SourceScope;
 import io.spine.tools.gradle.SpinePlugin;
 import io.spine.tools.gradle.TaskName;
-import io.spine.tools.dart.fs.SourceFile;
+import io.spine.tools.dart.fs.DartFile;
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
@@ -123,7 +123,7 @@ public final class McDartPlugin extends SpinePlugin {
 
     private void resolveImports(File sourceFile, Extension extension) {
         _debug().log("Resolving imports in file %s", sourceFile);
-        SourceFile file = SourceFile.read(sourceFile.toPath());
+        DartFile file = DartFile.read(sourceFile.toPath());
         Path libPath = extension.getLibDir()
                                 .getAsFile()
                                 .map(File::toPath)
