+++ tools/mc-js/src/main/java/io/spine/tools/mc/js/code/index/TypeParsers.java
@@ -30,9 +30,9 @@ import com.google.common.collect.Maps;
 import com.google.protobuf.Descriptors.FileDescriptor;
 import io.spine.tools.js.code.TypeName;
 import io.spine.code.proto.FileSet;
-import io.spine.tools.mc.js.code.snippet.Snippet;
+import io.spine.tools.mc.js.code.text.Snippet;
 import io.spine.tools.mc.js.code.CodeWriter;
-import io.spine.tools.mc.js.code.snippet.MapExport;
+import io.spine.tools.mc.js.code.text.MapExport;
 import io.spine.type.MessageType;
 import io.spine.type.TypeUrl;
