--- tools/mc-js/src/main/java/io/spine/tools/mc/js/code/snippet/Return.java
@@ -1,68 +1,0 @@
-/*
- * Copyright 2021, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.tools.mc.js.code.snippet;
-
-import com.google.errorprone.annotations.Immutable;
-import io.spine.tools.code.Line;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-import static java.lang.String.format;
-
-/**
- * A code line representing a {@code return} statement.
- */
-@Immutable
-public final class Return extends Line {
-
-    private Return(Object returnedValue) {
-        super(format("return %s;", returnedValue));
-    }
-
-    /**
-     * Composes a statement returning the value.
-     */
-    public static Return value(Object value) {
-        checkNotNull(value);
-        return new Return(value);
-    }
-
-    /**
-     * Composes a statement returning a string literal.
-     */
-    public static Return stringLiteral(String literal) {
-        checkNotNull(literal);
-        String quoted = format("'%s'", literal);
-        return new Return(quoted);
-    }
-
-    /**
-     * Composes a statement returning {@code null}.
-     */
-    public static Return nullReference() {
-        return value("null");
-    }
-}
