+++ tools/mc-js/src/main/java/io/spine/tools/mc/js/code/text/Comment.java
@@ -24,25 +24,38 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.js.code.imports;
+package io.spine.tools.mc.js.code.text;
 
-import java.io.File;
+import com.google.common.annotations.VisibleForTesting;
+import com.google.errorprone.annotations.Immutable;
+import io.spine.tools.code.Line;
 
-import static java.lang.String.format;
+import static com.google.common.base.Preconditions.checkNotNull;
 
-final class Given {
+/**
+ * A single-line comment.
+ */
+@Immutable
+public final class Comment extends Line {
 
-    /** Prevents instantiation of this utility class. */
-    private Given() {
+    @VisibleForTesting
+    static final String MC_JS_SIGNATURE = "Generated by Spine Model Compiler for JS";
+
+    private Comment(String text) {
+        super("// " + checkNotNull(text));
     }
 
-    static ImportStatement importWithPath(String path, File importOrigin) {
-        String importText = format("let foo = require('%s');", path);
-        JsFile file = new JsFile(importOrigin.toPath());
-        return new ImportStatement(file, importText);
+    /**
+     * Creates a comment code line from the passed text.
+     */
+    public static Comment of(String text) {
+        return new Comment(text);
     }
 
-    static String relativeImportPath() {
-        return "../path-relative-to-parent.js";
+    /**
+     * Obtains the comment to mark the generated code.
+     */
+    public static Comment generatedBySpine() {
+        return new Comment(MC_JS_SIGNATURE);
     }
 }
