+++ tools/mc-js/src/main/java/io/spine/tools/mc/js/code/text/Import.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.js.code.imports;
+package io.spine.tools.mc.js.code.text;
 
 import com.google.errorprone.annotations.Immutable;
 import io.spine.tools.code.Line;
