+++ tools/mc-js/src/main/java/io/spine/tools/mc/js/code/text/Method.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.js.code.snippet;
+package io.spine.tools.mc.js.code.text;
 
 import io.spine.tools.js.code.MethodReference;
 import io.spine.tools.code.Line;
