+++ tools/mc-js/src/main/java/io/spine/tools/mc/js/code/text/Snippet.java
@@ -24,38 +24,19 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.js.code.snippet;
+package io.spine.tools.mc.js.code.text;
 
-import com.google.common.annotations.VisibleForTesting;
-import com.google.errorprone.annotations.Immutable;
-import io.spine.tools.code.Line;
-
-import static com.google.common.base.Preconditions.checkNotNull;
+import io.spine.tools.mc.js.code.CodeWriter;
 
 /**
- * A single-line comment.
+ * A snippet of the code.
  */
-@Immutable
-public final class Comment extends Line {
-
-    @VisibleForTesting
-    static final String MC_JS_SIGNATURE = "Generated by Spine Model Compiler for JS";
-
-    private Comment(String text) {
-        super("// " + checkNotNull(text));
-    }
+public interface Snippet {
 
     /**
-     * Creates a comment code line from the passed text.
+     * Obtains code lines representing this snippet.
+     *
+     * @return always returns a new {@link CodeWriter}
      */
-    public static Comment of(String text) {
-        return new Comment(text);
-    }
-
-    /**
-     * Obtains the comment to mark the generated code.
-     */
-    public static Comment generatedBySpine() {
-        return new Comment(MC_JS_SIGNATURE);
-    }
+    CodeWriter writer();
 }
