+++ tools/mc-js/src/test/java/io/spine/tools/mc/js/code/text/ImportTest.java
@@ -24,11 +24,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.js.code.snippet;
+package io.spine.tools.mc.js.code.text;
 
 import com.google.protobuf.Any;
 import io.spine.tools.js.fs.FileName;
-import io.spine.tools.mc.js.code.imports.Import;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
