+++ tools/mc-js/src/test/java/io/spine/tools/mc/js/code/text/MethodTest.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.js.code.snippet;
+package io.spine.tools.mc.js.code.text;
 
 import com.google.common.truth.StringSubject;
 import com.google.common.truth.Truth;
