+++ tools/mc-js/src/test/java/io/spine/tools/mc/js/fs/JsFileTest.java
@@ -26,7 +26,6 @@
 
 package io.spine.tools.mc.js.fs;
 
-import io.spine.tools.mc.js.code.imports.JsFile;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
